first version
Cover Letter

We are pleased to submit our manuscript entitled "A Stream Order Family and Order-Based Parallel River Network Routing Method" for consideration for publication in Water.

Our research introduces a novel stream ordering method and an order-based parallel routing method, which serve as a significant advancement over the layer-based parallel routing method employed in the LisFlood model. Our approach has demonstrated improved performance, with enhanced efficiency and scalability. These attributes are particularly crucial for the timely forecasting of floods, a topic that aligns well with the central themes of the Water journal.

We have conducted a thorough evaluation and comparison with existing models, such as mizuRoute and LisFlood. Our methods have shown superior performance in terms of parallel efficiency and scalability. This advancement is expected to contribute significantly to the field of hydrology and flood modeling.

We confirm that neither the manuscript nor any parts of its content are currently under consideration for publication with or published in another journal.

All authors have approved the manuscript and agree with its submission to Water.

Thank you for considering our work for publication. We look forward to your positive response.